Extract mappings into a file
