+++ packages/site/src/organization-manager.ts
@@ -1,13 +1,11 @@
 import { ProjectStateEnum, SpheronApi } from "@spheron/core";
-import {
-  Organization,
-  Project,
-  UsageWithLimits,
-  mapCoreOrganization,
-  mapCoreProject,
-  mapCoreUsageWithLimits,
-} from "./interfaces";
+import { Organization, Project, UsageWithLimits } from "./interfaces";
 import OrganizationIdExtractor from "./organizationId-extractor";
+import {
+  mapCoreProject,
+  mapCoreOrganization,
+  mapCoreUsageWithLimits,
+} from "./interface-mappers";
 
 class OrganizationManager extends OrganizationIdExtractor {
   private readonly spheronApi: SpheronApi;
